refactor(issue #1): reorganize project into Clean Architecture layers
- Moved http/controllers/, http/views/, http/router.ts, and http/server.ts to interface/ folder.
- Moved database/database.ts, database/migrations/, and database/models/ to infrastructure/ folder.
