+++ src/infrastructure/database/database.ts
@@ -1,5 +1,5 @@
 import Knex from 'knex';
-import { config } from '../config.js';
+import { config } from '../../config.js';
 import { BaseModel } from './models/BaseModel.js';
 
 const makeDatabase = () => {
