+++ src/infrastructure/database/models/BaseModel.ts
