+++ src/infrastructure/database/models/DealershipModel.ts
