+++ src/infrastructure/database/models/VehicleModel.ts
