+++ src/interface/http/controllers/usersController.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
-import { handler } from '../../_lib/http/handler.js';
-import { UserModel } from '../../database/models/UserModel.js';
+import { handler } from '../../../_lib/http/handler.js';
+import { UserModel } from '../../../infrastructure/database/models/UserModel.js';
 
 const index = handler(async (_request, reply) => {
   const users = await UserModel.query();
