+++ src/interface/http/views/vehicles/_vehicle.ejs
